{"types":"C","runs":50}

+++ problem15.java
@@ -5,7 +5,7 @@ public class problem15 extends ConsoleProgram {
 	public void run() {
 		double firstNum = (int) (readInt("First number is: "));
 		double secondNum = (int) (readInt("Second number is: "));
-		double average = (firstNum + secondNum) / 2;
+		int average = (firstNum + secondNum) / 2;
 		println("Arithmetic average is: " + average);		
 	}
 }
